Extract MockCloud into mock_cloud_test.go and remove deallocTestCloud
Move MockCloud, MockInstance, cloudImpl, and MockCloudSetup from
taskrun_helpers_test.go into a dedicated mock_cloud_test.go file.

Merge deallocTestCloud capabilities (FailListInstances, FailTerminate,
TerminatedIDs) into MockCloud so a single mock serves both integration
and unit tests. Remove the redundant deallocTestCloud type.

Made-with: Cursor
